refactor(views/search): Reestruturar SearchScreen em componentes menores
- Move o SearchScreen.vue para a pasta SearchScreen/.
- Divide a lógica e o template do SearchScreen em SearchResultsRestaurants e SearchResultsProducts para maior clareza e aderência ao SRP.
- Prepara a nova estrutura para o template principal.
